refactor(memory): split into frontmatter/recall/actions modules
Brings packages/crane-mcp/src/tools/memory.ts under structural caps
(max-lines: 500, max-lines-per-function: 75, complexity: 15).

- memory.ts (entry, 189 LOC): thin orchestrator + re-exports
- memory-frontmatter.ts: types, YAML parse, body extraction, validation,
  serialization. serializeFrontmatter complexity reduced from 23 to <15
  by extracting serializeAppliesWhen + serializeSupersedes.
- memory-recall.ts: memoryability checks, glob matching, scoring, FTS5
  query mode, context mode. scoreMemory reduced from 16 to <15.
  buildUpdatedFm split into mergeBaseFields/mergeSeverity/mergeAppliesWhen/
  mergeArrayFields helpers. no-misused-promises fixed via void+catch.
- memory-actions.ts: per-action handlers (save/list/get/update/deprecate/
  recall) and shared helpers.

executeMemory complexity dropped from 90 to trivial (if/return dispatch).
Public API unchanged. 53/53 tests pass.
